feat: add discounts cart codes functional
