+++ src/features/cart/ui/cartProductList/CartProductList.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../../app/styles/variables.scss";
+@import "../../../../app/styles/variables.scss";
 
 .productList {
   display: flex;
